Fix compilation warning: add explicit cast for Math.pow
- Added explicit cast from double to int for Math.pow result
- Resolves 'implicit cast is possibly lossy' compiler warning
- Project uses -Werror flag which treats warnings as errors

+++ src/main/java/com/thealgorithms/maths/ArmstrongNumber.java
@@ -28,7 +28,7 @@ public final class ArmstrongNumber {
 
         while (number > 0) {
             int digit = number % 10;
-            sum += Math.pow(digit, digits);
+	    sum += (int) Math.pow(digit, digits);
             number /= 10;
         }
